refactor(mcp): drain screen_* view-command tool domain into a scoped pack (EP-8DC217EB BET-4)
Relocates the eleven Pseudo-User Contract screen_* tools (describe,
get_state, select_entity, navigate, open/close_panel, focus_field,
set_input, scroll_to, propose_action, dispatch_action) out of the
mcp-tools.ts executeTool switch into apps/web/lib/mcp/packs/screen-pack.ts.

Definitions moved verbatim (provenance-free descriptions); each handler
reproduces the former switch case exactly, resolving prisma,
getErrorMessage, findManifestForRoute, envelope-actions, and the recursive
executeTool dispatch through shared module imports so the pack carries no
static dependency on mcp-tools.ts internals. Grants mirror agent-grants.ts
TOOL_TO_GRANTS (the gating source stays authoritative). Pure relocation —
the tools behave identically via MCP; the existing mcp-tools.screen.test.ts
that drives them through executeTool stays green, and a new screen-pack.test.ts
asserts registration, grant mirror, and per-handler envelopes.

mcp-tools.ts: 16260 → 15647 LOC. Tool-pack ratchet baseline shrunk
227 → 216 (the eleven screen_* names removed); module-size baseline
retightened.

Process-Spine-Decision: EP-8DC217EB plan §4 BET-4 — drain the screen_* tool domain into a scoped pack; ratchet baseline shrunk; behavior-preserving relocation.
