+++ scripts/mcp-tool-pack-baseline.json
@@ -167,17 +167,6 @@
   "save_phase_handoff",
   "schedule_promotion",
   "schedule_release_bundle",
-  "screen_close_panel",
-  "screen_describe",
-  "screen_dispatch_action",
-  "screen_focus_field",
-  "screen_get_state",
-  "screen_navigate",
-  "screen_open_panel",
-  "screen_propose_action",
-  "screen_scroll_to",
-  "screen_select_entity",
-  "screen_set_input",
   "search_code_graph",
   "search_design_intelligence",
   "search_knowledge",
